requirements.txt
requests>=2.28.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.0.0
uvicorn>=0.20.0
fastapi>=0.100.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.24.0
websockets>=11.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0